refactor: add modules directory
