+++ api/modules/todo/todo.controller.ts
@@ -10,17 +10,19 @@ interface ITodoDetailRequest {
 
 interface ICreateRequest {
   Body: {
-    id: string;
     title: string;
   };
 }
 
 export class TodoController {
+  private todoService: TodoService;
+  constructor() {
+    this.todoService = new TodoService();
+  }
   // todos list
   public getTodosListHandler(request: FastifyRequest, reply: FastifyReply) {
-    const todoService = new TodoService();
     reply.header("Content-Type", "application/json").code(200);
-    reply.send(todoService.getTodosList());
+    reply.send(this.todoService.getTodosList());
   }
 
   // lookup todo by id
@@ -28,11 +30,12 @@ export class TodoController {
     request: FastifyRequest<ITodoDetailRequest>,
     reply: FastifyReply
   ) {
-    const todoService = new TodoService();
-
     reply.header("Content-Type", "application/json");
+
     const { id } = request.params;
-    const todo: Todo | undefined = todoService.getTodo(Number.parseInt(id));
+    const todo: Todo | undefined = this.todoService.getTodo(
+      Number.parseInt(id)
+    );
     if (todo === undefined) {
       reply.code(404).send({ message: "todo not found" });
     } else {
@@ -44,11 +47,9 @@ export class TodoController {
     request: FastifyRequest<ICreateRequest>,
     reply: FastifyReply
   ) {
-    const todoService = new TodoService();
-
     reply.header("Content-Type", "application/json");
     const { title } = request.body;
-    const todo = await todoService.createTodo(title);
+    const todo = await this.todoService.createTodo(title);
     reply.code(200).send(todo);
   }
 }
