+++ api/router.ts
@@ -1,5 +1,5 @@
 import { FastifyInstance } from "fastify";
-import { TodoRoutes } from "./todo/todo.routes";
+import { TodoRoutes } from "./modules/todo/todo.routes";
 
 // ルーティングをまとめてくれる関数
 export function Router(server: FastifyInstance): void {
